third_party/blink/web_tests/media/media-fragments/TC0067-expected.txt


Title: TC0067
Fragment: 't=00:00:00s'
Comment: UA knows that this is an invalid media fragment, so it will play the entire media resource.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
END OF TEST
